word translation

+++ src/config.js
@@ -1,31 +1,18 @@
-
 const configPath = "$sandbox/config.json";
-const HelpUrl = "https://github.com/solywsh/bob-relingo"
-
-const supportedLanguages = [
-    ['auto', 'auto'],
-    ['zh-Hans', 'zh-Hans'],
-    ['zh-Hant', 'zh-Hant'],
-    ['en', 'en'],
-];
-
-function getConfig() {
-    configInit();
-    var config = $file.read(configPath).toUTF8();
-    $log.info(config);
-    return JSON.parse(config);
-}
-
+const relingoConfigPath = "$sandbox/relingoConfig.json";
 
 const defaultConfig = {
+    "version": "0.0.1", // 配置版本号与插件版本区分
     "username": "", // 用户名
-    "mail": $option.email, // 邮箱
+    "mail": "", // 邮箱
     "mailCode": "", // 邮箱验证码
     "mailCodeSendAt": 0, // 邮箱验证码发送时间
     "token": "", // token
     "tokenUpdatedAt": 0, // token更新时间new Date().getTime()
+    "tokenExpireAt": 0, // token过期时间
     "status": "init", // 状态 init waitingCode active error
-    "lang": "cn", //
+    "lang": "cn", // header请求头用，暂时不知道有什么用
+    "native": "zh", // 目标翻译语言
     "strange": "", // 陌生单词本
     "mastered": "", // 掌握的单词本
     books: [], // relingo books
@@ -34,27 +21,43 @@ const defaultConfig = {
 function updateConfig(config) {
     $file.write({
         data: $data.fromUTF8(JSON.stringify(config)),
-        path: config
+        path: configPath,
     })
 }
 
-function configInit(){
+function updateRelingoConfig(config) {
+    $file.write({
+        data: $data.fromUTF8(config),
+        path: relingoConfigPath,
+    })
+}
+
+function getConfig() {
+    configInit();
+    const c = JSON.parse($file.read(configPath).toUTF8());
+    if (c.vserion < defaultConfig.vserion) {
+        $file.delete(configPath);
+        configInit();
+    }
+    // $log.info(configData);
+    return JSON.parse($file.read(configPath).toUTF8());
+}
+function configInit() {
     if ($file.exists(configPath) && !$file.isDirectory(configPath)) {
-        $log.info("config file exit / is not directory");
+        // $log.info("config file exit / is not directory");
         return;
     }
-    $file.delete(configPath);
-
+    if ($file.exists(configPath) && $file.isDirectory(configPath)) {
+        $file.delete(configPath);
+    }
+    // $log.info(JSON.stringify(defaultConfig));
     $file.write({
         data: $data.fromUTF8(JSON.stringify(defaultConfig)),
         path: configPath
     })
 }
 
-exports.supportedLanguages = supportedLanguages;
-
 exports.configPath = configPath;
-exports.helpUrl = HelpUrl;
-
 exports.getConfig = getConfig;
 exports.updateConfig = updateConfig;
+exports.updateRelingoConfig = updateRelingoConfig;
